NLA SoC: Graph Editor Menus + Operator Name Cleanup
As with the DopeSheet, the names of operators in the Graph Editor have been cleaned up, and operators have been added to menus as appropriate to show their availability.

Tweaked a few DopeSheet operator names to be more in line with the Graph Editor equivalents, and vica versa.


TODO: now, the operator poll callbacks here need checking...